Move out Wrap, WrapAs and ToErk from error.go
